requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.28.0
nltk>=3.8
numpy>=1.24.0
rank_bm25>=0.2.2
scikit-learn>=1.3.0
